fix layout worker in background thread, and fix jest
