+++ src/data/graphs/Gallery_Demos/demo_SceneGraph_e8petrieProjection.ts
@@ -3,7 +3,7 @@ import { SceneGraph } from "../../../core/model/SceneGraph";
 // Import the SVG as a raw string
 import petrie421t2b6SvgUrl from "!!raw-loader!/public/svgs/e8_4_21_t2-B6.svg";
 import petriesvgUrl from "!!raw-loader!/public/svgs/e8petrie.svg";
-import { PresetLayoutType } from "../../../core/layouts/LayoutEngine";
+import { PresetLayoutType } from "../../../core/layouts/layoutEngineTypes";
 import { loadSvgToSceneGraph } from "../../../utils/svgLoader";
 
 /**
